web_form/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
black==23.10.0
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.7
Django==4.2
django-bootstrap3==23.1
django-filter==23.1
django-tables2==2.5.3
djlint==1.24.0
EditorConfig==0.12.3
et-xmlfile==1.1.0
flake8==6.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
isort==5.12.0
jsbeautifier==1.14.7
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.24.3
openpyxl==3.1.2
packaging==23.2
pandas==2.0.1
pathspec==0.11.1
pep8==1.7.1
platformdirs==3.11.0
psycopg2-binary==2.9.3
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.4
tablib==3.4.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.8.0
tzdata==2023.3
